+++ tests/TestCase/Executor/PostgresExecutorTest.php
@@ -13,17 +13,17 @@ declare(strict_types=1);
  * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
 
-namespace DatabaseBackup\Test\TestCase\Driver;
+namespace DatabaseBackup\Test\TestCase\Executor;
 
-use DatabaseBackup\Driver\Postgres;
+use DatabaseBackup\Executor\PostgresExecutor;
 use DatabaseBackup\TestSuite\DriverTestCase;
 use PHPUnit\Framework\Attributes\CoversClass;
 
 /**
- * PostgresTest class.
+ * PostgresExecutorTest class.
  */
-#[CoversClass(Postgres::class)]
-class PostgresTest extends DriverTestCase
+#[CoversClass(PostgresExecutor::class)]
+class PostgresExecutorTest extends DriverTestCase
 {
     /**
      * @inheritDoc
@@ -32,8 +32,8 @@ class PostgresTest extends DriverTestCase
     {
         parent::setUp();
 
-        if ($this->getConnection()->config()['scheme'] !== 'postgres') {
-            $this->markTestSkipped('Skipping tests for postgres, current driver is `' . $this->getConnection()->config()['scheme'] . '`');
+        if (!$this->Executor instanceof PostgresExecutor) {
+            $this->markTestSkipped('Skipping tests for `PostgresExecutor`, current driver is `' . $this->Executor::class . '`');
         }
     }
 }