+++ CHANGELOG.md
@@ -5,6 +5,13 @@
   property via the magic methods `getX()` (be careful not to confuse the `getCompression()` method, which now returns
   the `$compression` property of `BackupExport`, with the old method provided by `BackupTrait`, which had been
   deprecated and has now been removed);
+* class `DatabaseBackup\Driver\AbstractDriver` has become `DatabaseBackup\Executor\AbstractExecutor`, class
+  `DatabaseBackup\Driver\Mysql` has become `DatabaseBackup\Executor\MysqlExecutor`, class `DatabaseBackup\Driver\Postgres`
+  has become `DatabaseBackup\Executor\PostgresExecutor` and class `DatabaseBackup\Driver\Sqlite` has become 
+  `DatabaseBackup\Executor\SqliteExecutor`. Aliases have been added to old classes for backwards compatibility, but will
+  be removed in a future release;
+* added new `AbstractBackupUtility::getExecutor()` that gets the `Executor` instance (old `Driver` classes now renamed)
+  according to the connection; 
 * passing the `$compression` argument as a string or `null` to `BackupExport::compression()` had been deprecated and has
   been removed (backwards compatibility removed);
 * the `BackupManager::index` methods no longer returns, in the array for each file, the `filename` key;
@@ -16,8 +23,12 @@
 * the `BackupImport::import()` methods now throws a `BadMethodCallException` exception (rather than a `LogicException`)
   when the filename has not been set;
 * the `TestCase::createSomeBackups()` method has been improved;
+* the `AbstractBackupUtility::getDriver()` method is deprecated and will be removed in a future release. Use instead the
+  `getExecutor()` method;
 * the `AbstractBackupUtility::__get()` method is deprecated and will be removed in a future release;
 * `DeleteAllCommand`, `RotateCommand` and `SendCommand` classes had been deprecated and have been removed;
+* the `getConnection()` and `getDriverName()` methods provided by `BackupTrait` are deprecated and will be removed in a
+  future release. The entire `BackupTrait` trait is now deprecated and will be removed in a future release;
 * `getAbsolutePath()`, `getCompression()`, `getExtension()` and `getValidCompressions()` methods provided by 
   `BackupTrait` had been deprecated and has been removed;
 * all classes, methods and code related to sending backups via email had been deprecated, and now they have been
@@ -389,7 +400,7 @@
 ### 2.1.1
 * `afterExport()`, `afterImport()`, `beforeExport()` and `beforeImport` methods
     are now real events;
-* now you can choose if you want to redirects stderr to `/dev/null`. This
+* now you can choose if you want to redirect stderr to `/dev/null`. This
     suppresses the output of executed commands.
 
 ### 2.1.0
@@ -450,7 +461,7 @@
 
 ### 1.0.1
 * added `BackupManager::deleteAll()` and `BackupShell::deleteAll()` methods;
-* improved tests. Also errors in the shell are checked.
+* improved tests. Also, errors in the shell are checked.
 
 ### 1.0.0
 * first release.